# frontend-gestion-de-stock/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation for Gestion de stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Lotfi FATHI",
      "url" : "https://lotfif.github.io/portfolio/",
      "email" : "lotfi.fathi@gmail.com"
    }
  },
  "basePath" : "http://localhost:8081",
...
  "paths" : {
...
    "/gestiondestock/v0/articles/{idArticle}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Recherche un article par ID",
        "description" : "Cette méthode permet de chercher et récupérer un article par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "l'article a été trouvé dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDto"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v0/categories/{idCategory}" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Recherche une categorie par ID",
        "description" : "Cette méthode permet de chercher et récupérer une categorie par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCategory",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la catégorie a été trouvé dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CategoryDto"
            }
          },
          "404" : {
            "description" : "Aucune catégorie n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v0/clients/{idClient}" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "Recherche un client par ID",
        "description" : "Cette méthode permet de chercher et récupérer un client par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "le client a été trouvé dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ClientDto"
            }
          },
          "404" : {
            "description" : "Aucun client n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v0/commandesclients/{idCommandeClient}" : {
      "get" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/commandesfournisseurs/update/fournisseur/{idCommande}/{idFournisseur}" : {
      "patch" : {
        "tags" : [ "commandesfournisseurs" ],
        "operationId" : "updateFournisseur",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/commandesfournisseurs/{idCommandeFournisseur}" : {
      "get" : {
        "tags" : [ "commandesfournisseurs" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/entreprises/{idEntreprise}" : {
      "get" : {
        "tags" : [ "entreprises" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idEntreprise",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/EntrepriseDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/fournisseurs/create" : {
      "post" : {
        "tags" : [ "fournisseurs" ],
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FournisseurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/fournisseurs/delete/{idFournisseur}" : {
      "delete" : {
        "tags" : [ "fournisseurs" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/gestiondestock/v0/fournisseurs/{idFournisseur}" : {
      "get" : {
        "tags" : [ "fournisseurs" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/correctionneg" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "correctionStockNeg",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/correctionpos" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "correctionStockPos",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/entree" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "entreeStock",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/filter/article/{idArticle}" : {
      "get" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "mvtStkArticle",
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MvtStkDto"
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/sortie" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "sortieStock",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/mvtStk/stockreel/{idArticle}" : {
      "get" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "stockReelArticle",
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "number"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/save/{id}/{title}/{context}" : {
      "post" : {
        "tags" : [ "photos" ],
        "operationId" : "savePhoto",
        "parameters" : [ {
          "name" : "context",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "file",
          "in" : "formData",
          "required" : true,
          "type" : "file"
        }, {
          "name" : "title",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/utilisateurs/create" : {
      "post" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UtilisateurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/utilisateurs/delete/{idUtilisateur}" : {
      "delete" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/gestiondestock/v0/utilisateurs/find/{mail}" : {
      "get" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "findByMail",
        "parameters" : [ {
          "name" : "mail",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/utilisateurs/update/password" : {
      "post" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "changerMotDePasse",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ChangerMotDePasseUtilisateurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v0/utilisateurs/{idUtilisateur}" : {
      "get" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v0/ventes/{idVente}" : {
      "get" : {
        "tags" : [ "ventes" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idVente",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VenteDto"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "AdresseDto" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "ArticleDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "codeArticle" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "photo" : {
          "type" : "string"
        },
        "category" : {
          "$ref" : "#/definitions/CategoryDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "CategoryDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ChangerMotDePasseUtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "motDePasse" : {
          "type" : "string"
        },
        "confirmMotDePasse" : {
          "type" : "string"
        }
      }
    },
    "ClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "CommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "client" : {
          "$ref" : "#/definitions/ClientDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeClients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClientDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "CommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "fournisseur" : {
          "$ref" : "#/definitions/FournisseurDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseurDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "EntrepriseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "codeFiscal" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "siteWeb" : {
          "type" : "string"
        }
      }
    },
    "FournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneVenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "vente" : {
          "$ref" : "#/definitions/VenteDto"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "MvtStkDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateMvt" : {
          "type" : "integer",
          "format" : "int64"
        },
        "quantite" : {
          "type" : "number"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "typeMvt" : {
          "type" : "string",
          "enum" : [ "ENTREE", "SORTIE", "CORRECTION_POS", "CORRECTION_NEG" ]
        },
        "sourceMvt" : {
          "type" : "string",
          "enum" : [ "COMMANDE_CLIENT", "COMMANDE_FOURNISSEUR", "VENTE" ]
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "RolesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "roleName" : {
          "type" : "string"
        }
      }
    },
    "UtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "string",
          "format" : "date-time"
        },
        "motDePasse" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "entreprise" : {
          "$ref" : "#/definitions/EntrepriseDto"
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RolesDto"
          }
        }
      }
    },
    "VenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateVente" : {
          "type" : "integer",
          "format" : "int64"
        },
        "commentaire" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneVente" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVenteDto"
          }
        }
      }
    }
  }
}